requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
cffi==1.17.0
channels==4.1.0
channels-redis==4.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
Django==5.1
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
inflection==0.5.1
kombu==5.4.0
msgpack==1.0.8
numpy==2.0.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pillow==10.4.0
prompt_toolkit==3.0.47
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
redis==5.0.8
service-identity==24.1.0
setuptools==70.0.0
six==1.16.0
sqlparse==0.5.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
ucimlrepo==0.0.7
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.0.3
